+++ src/OnboardingSPA/components/Heading/AIHeading/stylesheet.scss
