+++ src/OnboardingSPA/components/Heading/HeadingWithSubHeading/stylesheet.scss
