+++ src/OnboardingSPA/steps/BasicInfo/index.js
@@ -1,5 +1,5 @@
 import CommonLayout from '../../components/Layouts/Common';
-import HeadingWithSubHeading from '../../components/HeadingWithSubHeading';
+import HeadingWithSubHeading from '../../components/Heading/HeadingWithSubHeading';
 import BasicInfoForm from './basicInfoForm';
 import { SIDEBAR_LEARN_MORE, VIEW_NAV_PRIMARY } from '../../../constants';
 import { store as nfdOnboardingStore } from '../../store';
