+++ src/OnboardingSPA/steps/DesignThemeStyles/Menu/index.js
@@ -5,7 +5,7 @@ import { useNavigate, useLocation } from 'react-router-dom';
 import getContents from '../contents';
 import { store as nfdOnboardingStore } from '../../../store';
 import CommonLayout from '../../../components/Layouts/Common';
-import HeadingWithSubHeading from '../../../components/HeadingWithSubHeading';
+import HeadingWithSubHeading from '../../../components/Heading/HeadingWithSubHeading';
 import { useGlobalStylesOutput } from '../../../utils/global-styles/use-global-styles-output';
 import { getPatterns } from '../../../utils/api/patterns';
 import { getGlobalStyles } from '../../../utils/api/themes';
