+++ src/OnboardingSPA/steps/SiteFeatures/index.js
@@ -8,7 +8,7 @@ import { store as nfdOnboardingStore } from '../../store';
 import { SIDEBAR_LEARN_MORE, VIEW_NAV_PRIMARY } from '../../../constants';
 import CommonLayout from '../../components/Layouts/Common';
 import { getSiteFeatures } from '../../utils/api/plugins';
-import HeadingWithSubHeading from '../../components/HeadingWithSubHeading';
+import HeadingWithSubHeading from '../../components/Heading/HeadingWithSubHeading';
 import CheckboxList from '../../components/CheckboxTemplate/CheckboxList';
 import { CheckboxListSkeleton } from '../../components/CheckboxTemplate';
 import getContents from './contents';
